Video Subtitle/requirements.txt
ffmpeg-python
faster-whisper
pandas
